% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{aExprsDataQC}
\alias{aExprsDataQC}
\title{Function for automatized quality control.}
\usage{
aExprsDataQC(demoData, vFlags)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object,
storing all expression, annotation, and probe information linked together.}

\item{vFlags}{Character string to set the flags for quality control.}
}
\value{
A matrix.
}
\description{
Function for automatized quality control.
}
\examples{
demoData <- file
vFlags <- "QCFlags"
}
