% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{build_matrix}
\alias{build_matrix}
\title{Builds a matrix from an input csv file, with one column as rownames}
\usage{
build_matrix(counts_data, x)
}
\arguments{
\item{counts_data}{A data frame with floats or integers that
has strings in one column.}

\item{x}{Number of column which contains the rownames description.}
}
\value{
A matrix.
}
\description{
Builds a matrix from an input csv file, with one column as rownames
}
\examples{
x <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
  package = "DgeaHeatmap"
))
matrixCounts <- build_matrix(input_data, x)
}
