% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{color_setting}
\alias{color_setting}
\title{Function to set color scheme for a heatmap.}
\usage{
color_setting(colorPalette)
}
\arguments{
\item{colorPalette}{Chosen color palette.}
}
\value{
No return.
}
\description{
Function to set color scheme for a heatmap.
}
\examples{
colorPalette <- "RdBu"
}
