% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{DGEAedgeR}
\alias{DGEAedgeR}
\title{DGEAedgeR}
\usage{
DGEAedgeR(
  rawCounts,
  metadata,
  grouping_columns,
  comparisons = NULL,
  prefix = "DEA"
)
}
\arguments{
\item{rawCounts}{A matrix containing raw counts (integers).}

\item{metadata}{A dataframe containing the meta data information of the raw
counts matrix.}

\item{grouping_columns}{A list of the columns names in the metadata file that
are used to generate groups for the DEA that are compared with each other.}

\item{comparisons}{An element of the class list, that contains optional
information on specific comparisons that are to be conducted (simplifies the
results).}

\item{prefix}{A string used as a prefix for the contrast_names that are
generated.}
}
\value{
A list containing the voom data, fitting of the linear model, results
of the contrasts, and the generated design.
}
\description{
Performs a differential expression analysis based on edgeR
}
\examples{
set.seed(1)
counts <- matrix(rnbinom(600, mu = 10, size = 1), ncol = 6)
colnames(counts) <- paste0("Sample", 1:6)
rownames(counts) <- paste0("Gene", 1:100)

# Create simple metadata
metadata <- data.frame(
    segment = factor(rep(c("A", "B"), each = 3)),
    row.names = colnames(counts)
)

comparisons <- list(
    Comp1 = c("A", "B")
)

results_edgeR <- DGEAedgeR(
    rawCounts = counts,
    metadata = metadata,
    grouping_columns = "segment",
    comparisons = comparisons,
    prefix = "DEA"
)

names(results_edgeR)
}
