% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Function to create an elbow plot to choose k for clustering by k-Means.}
\usage{
elbow_plot(top_genes_matrix, maxK = 15)
}
\arguments{
\item{top_genes_matrix}{A matrix for which the best number of clusters in
k-Means Clustering is supposed to be calculated.}

\item{maxK}{An integer determining the maximum number of possible clusters.}
}
\value{
An elbow plot to choose k.
}
\description{
Function to create an elbow plot to choose k for clustering by k-Means.
}
\examples{
x <- 1
seed <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
package = "DgeaHeatmap"))
matrixCounts <- build_matrix(input_data, x)
scaled_counts <- scale_counts(matrixCounts)
top_genes_matrix <- scaled_counts
set.seed(seed)
elbow_plot(top_genes_matrix)
}
