% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{extract_genes_direction}
\alias{extract_genes_direction}
\title{Get list of genes and their direction of regulation}
\usage{
extract_genes_direction(
  results_list,
  only_up = FALSE,
  only_down = FALSE,
  only_sig = FALSE,
  padj = NULL,
  padj_cutoff = 0.05,
  lfc_cutoff = 0
)
}
\arguments{
\item{results_list}{A list containing the results of the DEA.}

\item{only_up}{A Boolean opting for only up regulated genes.}

\item{only_down}{A Boolean opting for only down regulated genes.}

\item{only_sig}{A Boolean opting for only significantly regulated genes.}

\item{padj}{A Float, defining the adjusted p-value.}

\item{padj_cutoff}{A Float, setting the adjusted p-values cutoff.}

\item{lfc_cutoff}{A Float, setting the fold change cutoff.}
}
\value{
A character vector of contrast names.
}
\description{
Get list of genes and their direction of regulation
}
\examples{
rawCounts <- read.csv(
    system.file("extdata/RawDataExamplePackageNanostring.csv",
        package = "DgeaHeatmap"
    )
)
rawCounts <- build_matrix(rawCounts, 1)
metadata <- readRDS(system.file("extdata/MetaDataPackageNanostring.rds",
    package = "DgeaHeatmap"
))
grouping_columns <- c("segment", "region", "class", "slide_name")
comparisons <- list(
    Comp1 = c(
        "Geometric_Segment_glomerulus_DKD_disease3",
        "PanCK_tubule_DKD_disease4"
    ),
    Comp2 = c("neg_tubule_DKD_disease4", "PanCK_tubule_DKD_disease4")
)
storage.mode(rawCounts) <- "integer"
sum(!is.finite(as.matrix(rawCounts)))
results_list_DESeq2 <- DGEADESeq2(
    rawCounts, metadata, grouping_columns, comparisons
)
results_list <- results_list_DESeq2$results
gene_list <- extract_genes_direction(
    results_list,
    only_up = TRUE, only_down = FALSE, only_sig = FALSE,
    padj = NULL, padj_cutoff = 0.05, lfc_cutoff = 0
)

}
