% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{genRawReadCountTable}
\alias{genRawReadCountTable}
\title{Function to generating a raw read count table.}
\usage{
genRawReadCountTable(demoData)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object,
storing all expression, annotation, and probe information linked together.}
}
\value{
A dataframe containing only the raw read counts
that have passed quality control.
}
\description{
Function to generating a raw read count table.
}
\examples{
demoData <- readRDS(
  system.file(
    "extdata",
    "miniGeoMx.rds",
    package = "DgeaHeatmap"
  )
)

demoData <- add_demoElem(demoData)

NanoStringNCTools::assayDataApply(
  demoData,
  MARGIN = 1,
  FUN = mean,
  elt = "demoElem"
)[seq_len(5)]

demoData <- split_data_by_column(
  demoData,
  vGroup = "aoi",
  vElt = "demoElem"
)
df_Exp <- genRawReadCountTable(demoData)
}
