% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{get_dist}
\alias{get_dist}
\title{Calculates the distance matrix}
\usage{
get_dist(x, method)
}
\arguments{
\item{x}{An input matrix to create the distance matrix.}

\item{method}{Astring setting the distance method for clustering.}
}
\value{
A distance matrix.
}
\description{
Calculates the distance matrix
}
\examples{
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
  package = "DgeaHeatmap"
))
x <- 1
matrixCounts <- build_matrix(input_data, x)
matrix <- scale_counts(matrixCounts)
method <- "euclidean"
distance_matrix <- get_dist(matrix, method)
}
