% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{get_heatmap_colors}
\alias{get_heatmap_colors}
\title{Creating a color scheme based on the available color palettes of
RColorBrewer for the heatmap.}
\usage{
get_heatmap_colors(colorPalette)
}
\arguments{
\item{colorPalette}{An input string defining the color palette
(available from RColorBrewer).}
}
\value{
The colors used in heatmap based on either ComplexHeatmap
default or RColorBrewer.
}
\description{
Creating a color scheme based on the available color palettes of
RColorBrewer for the heatmap.
}
\examples{
colorPalette <- "RdBu"
get_heatmap_colors(colorPalette)
}
