% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{Kmean_generation}
\alias{Kmean_generation}
\title{Generates K-means for the columns and rows of a matrix.}
\usage{
Kmean_generation(m_top_genes_matrix, seed, k)
}
\arguments{
\item{m_top_genes_matrix}{An input matrix for which they K-Means are calculated next.}

\item{seed}{An integer to set the seed for the K-Mean calculation, allows reproducibility.}

\item{k}{Number k of how many clusters are created.}
}
\value{
All calculated K-Means of the matrix.
}
\description{
Generates K-means for the columns and rows of a matrix.
}
\examples{
m_top_genes_matrix <- matrix
seed <- 1
k <- 1
}
