% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{most_variable_genes}
\alias{most_variable_genes}
\title{Function to determine the most variable genes of each cluster to enable annotation..}
\usage{
most_variable_genes(m_kmeans, number_of_annotations_per_cluster, k)
}
\arguments{
\item{m_kmeans}{Matrix of the k-Means for each row in a gene set.}

\item{number_of_annotations_per_cluster}{Number of wanted annotations per cluster.}

\item{k}{Number of clusters.}
}
\value{
A list of the most variable rows (genes) of each cluster.
}
\description{
Function to determine the most variable genes of each cluster to enable annotation..
}
\examples{
m_kmeans <- matrix
number_of_annotations_per_cluster <- 10
k <- 1
}
