% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{Create pairwise contrasts}
\usage{
pairwise_contrasts(comparisons, comp_factor, prefix = "Contrast")
}
\arguments{
\item{comparisons}{An element of the class list, that contains optional
information on specific comparisons that are to be conducted (simplifies the
results).}

\item{comp_factor}{A factor, for fallback all pairwise comparison.}

\item{prefix}{A string used as a prefix for the contrast_names that are
generated.}
}
\value{
A character vector of contrast names.
}
\description{
Create pairwise contrasts
}
\examples{
rawCounts <- read.csv(system.file(
    "extdata/RawDataExamplePackageNanostring.csv",
    package = "DgeaHeatmap"
))
rawCounts <- build_matrix(rawCounts, 1)
metadata <- readRDS(system.file(
    "extdata/MetaDataPackageNanostring.rds",
    package = "DgeaHeatmap"
))
grouping_columns <- c("segment", "region", "class", "slide_name")
comparisons <- list(
    Comp1 = c(
        "Geometric_Segment_glomerulus_DKD_disease3",
        "PanCK_tubule_DKD_disease4"
    ),
    Comp2 = c("neg_tubule_DKD_disease4", "PanCK_tubule_DKD_disease4")
)
comp <- apply(
    metadata[, grouping_columns, drop = FALSE], 1, paste,
    collapse = "_"
)
comp <- gsub(" ", "_", comp) # Replace spaces with underscores
metadata$comp <- make.names(comp) # Clean names for use in model.matrix
comp_factor <- factor(metadata$comp)
paired_contrasts <- pairwise_contrasts(comparisons, comp_factor)

}
