% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{performing_kMeans}
\alias{performing_kMeans}
\title{Function to perform k-Means clustering for a matrix and setting split to split a heatmap into clusters.}
\usage{
performing_kMeans(m_top_genes_matrix, k)
}
\arguments{
\item{m_top_genes_matrix}{A matrix for which K-Means are calculated.}

\item{k}{Number of clusters for k-Means clustering (chosen by elbow-plot).}
}
\value{
A new dataframe containing information assigning each rows of the input matrix to a cluster.
}
\description{
Function to perform k-Means clustering for a matrix and setting split to split a heatmap into clusters.
}
\examples{
m_top_genes_matrix <- matrix
k <-- 3
}
