% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{print_heatmap}
\alias{print_heatmap}
\title{Function to build a heatmap using other functions.}
\usage{
print_heatmap(
  seed,
  m_top_genes_matrix,
  title,
  split,
  anno,
  fontsize_columnNames,
  fontsize_rowNames,
  title_heatmapLegend,
  WidthNum,
  HeightNum,
  UnitSize
)
}
\arguments{
\item{seed}{An integer used to set seed for k-Means clustering.}

\item{m_top_genes_matrix}{An input matrix for which the heatmap is created.}

\item{title}{A string used to set the title of the heatmap.}

\item{split}{A dataframe containing information about the clustering of the rows of the input matrix.}

\item{fontsize_columnNames}{An integer used to set the font size of the columns in the heatmap.}

\item{fontsize_rowNames}{An integer used to set the font size of the rownames in the heatmap.}

\item{title_heatmapLegend}{A string setting the title of the legend of the heatmap.}

\item{WidthNum}{A float setting the width of the heatmap.}

\item{HeightNum}{A float setting the height of the heatmap.}

\item{UnitSize}{A string such as "cm" or "inch" to set the unit of HeightNum and WidthNum.}
}
\value{
A plotted heatmap.
}
\description{
Function to build a heatmap using other functions.
}
\examples{
seed <- 1
m_top_genes_matrix <- matrix
title <- "Heatmap of Data"
split <- performing_kMeans(m_top_genes_matrix)
fontsize_columnNames <- 6
fontsize_rowNames <- 4
title_heatmaoLegend <- "Expression"
WidthNum <- 4.5
HeightNum <- 3
UnitSize <- "cm"
}
