% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{set_annotation}
\alias{set_annotation}
\title{Function to set annotation for a heatmap.}
\usage{
set_annotation(m_top_genes_matrix, top_x_genes_cluster)
}
\arguments{
\item{m_top_genes_matrix}{An input matrix to used for heatmap generation and choosing annotation for a heatmap.}

\item{top_x_genes_cluster}{A list of the most variable x genes of each cluster in a heatmap.}
}
\value{
A numeric index from the orginal matrix.
}
\description{
Function to set annotation for a heatmap.
}
