% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{show_data_distribution}
\alias{show_data_distribution}
\title{Function to visualize the data distribution within the data of a matrix.}
\usage{
show_data_distribution(scaled_counts)
}
\arguments{
\item{scaled_counts}{An input matrix with Z-count scaled values.}
}
\value{
A plot visualizing the data distribution of Z-count scaled data.
}
\description{
Function to visualize the data distribution within the data of a matrix.
}
\examples{
x <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
    package = "DgeaHeatmap"
))
matrixCounts <- build_matrix(input_data, x)
scaled_counts <- scale_counts(matrixCounts)
show_data_distribution(scaled_counts)
}
