% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{split_data_by_column}
\alias{split_data_by_column}
\title{Splitting data by group column with feature,
pheno or protocol data to then get the mean.}
\usage{
split_data_by_column(demoData, vGroup, vElt)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object,
storing all expression, annotation, and probe information linked together.}

\item{vGroup}{Group of features in column.}

\item{vElt}{Character string of the expression matrix name.}
}
\value{
A NanostringGeoMxSet object.
}
\description{
Splitting data by group column with feature,
pheno or protocol data to then get the mean.
}
\examples{
demoData <- readRDS(
  system.file(
    "extdata",
    "miniGeoMx.rds",
    package = "DgeaHeatmap"
  )
)

demoData <- add_demoElem(demoData)

NanoStringNCTools::assayDataApply(
  demoData,
  MARGIN = 1,
  FUN = mean,
  elt = "demoElem"
)[seq_len(5)]

demoData <- split_data_by_column(
  demoData,
  vGroup = "aoi",
  vElt = "demoElem"
)
}
