% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{add_demoElem}
\alias{add_demoElem}
\title{Build a matrix from an input csv file, with one column as rownames}
\usage{
add_demoElem(demoData)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object.}
}
\value{
A NanostringGeoMxSet object.
}
\description{
Build a matrix from an input csv file, with one column as rownames
}
\examples{
demoData <- readRDS(
    system.file(
        "extdata",
        "miniGeoMx.rds",
        package = "DgeaHeatmap"
    )
)

demoData <- add_demoElem(demoData)
}
