% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{add_demoElem}
\alias{add_demoElem}
\title{Build a matrix from an input csv file, with one column as rownames}
\usage{
add_demoElem(demoData)
}
\arguments{
\item{demoData}{A demoData object, storing all expression, annotation, and probe information linked together.}
}
\value{
A matrix.
}
\description{
Build a matrix from an input csv file, with one column as rownames
}
\examples{
demoData <- file
}
