% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{add_demoElem}
\alias{add_demoElem}
\title{Build a matrix from an input csv file, with one column as rownames}
\usage{
add_demoElem(demoData)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object.}
}
\value{
A NanostringGeoMxSet object.
}
\description{
Build a matrix from an input csv file, with one column as rownames
}
\examples{
\dontrun{
datadir <- system.file("extdata",
  "WTA_NGS_Example",
  package = "GeoMxWorkflows"
)
DCCFiles <- dir(file.path(datadir, "dccs"),
  pattern = ".dcc$", full.names = TRUE, recursive = TRUE
)
PKCFiles <- unzip(
  zipfile = dir(file.path(datadir, "pkcs"),
    pattern = ".zip$", full.names = TRUE, recursive = TRUE
  ),
  exdir = tempdir()
)
SampleAnnotationFile <- dir(file.path(datadir, "annotation"),
  pattern = ".xlsx$", full.names = TRUE, recursive = TRUE
)
demoData <- suppressWarnings(GeomxTools::readNanoStringGeoMxSet(
  dccFiles = DCCFiles, pkcFiles = PKCFiles,
  phenoDataFile = SampleAnnotationFile,
  phenoDataSheet = "Template", phenoDataDccColName = "Sample_ID",
  protocolDataColNames = c("aoi", "roi"), configFile = NULL,
  analyte = "RNA", phenoDataColPrefix = "", experimentDataColNames = NULL
))
demoData <- add_demoElem(demoData)
}
}
