% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{adv_Heatmap}
\alias{adv_Heatmap}
\title{Creating a color scheme based on the available color palettes of RColorBrewer for the heatmap.}
\usage{
adv_Heatmap(
  ncounts_matrix,
  seed = 1,
  column_name = "Heatmap",
  colorPalette = NULL,
  cluster_method = "hierarchical",
  distance_method = "euclidean",
  cluster_rows = TRUE,
  cluster_columns = FALSE,
  k_row = NULL,
  k_col = NULL,
  sample_metadata = NULL,
  annotation_colors = NULL,
  annotation_name_side = "right",
  show_row_names = FALSE,
  show_column_names = TRUE,
  row_annotation = FALSE,
  row_annotation_method = "auto",
  row_anno_names = NULL,
  row_anno_number = 5,
  fontsize_rowAnnotation = 8,
  fontsize_columnNames = 6,
  fontsize_rowNames = 4,
  title_heatmapLegend = "Expression",
  WidthNum = 4.5,
  HeightNum = 3,
  UnitSize = "cm"
)
}
\arguments{
\item{ncounts_matrix}{An input matrix to create the heatmap.}

\item{seed}{An integer used to set seed for reproducibility and k-Mean clustering.}

\item{column_name}{A string to set the title of the heatmap.}

\item{colorPalette}{Name of the colorPalette used for the heatmap.}

\item{cluster_method}{A string setting the cluster method for the heatmap.}

\item{distance_method}{A string setting the distance method for clustering.}

\item{cluster_rows}{A Boolean switching the optional clustering of the rows on and off.}

\item{cluster_columns}{A Boolean switching the optional clustering of the columns on and off.}

\item{k_row}{An integer used to set number of clusters for row clustering in the heatmap.}

\item{k_col}{An integer used to set number of clusters for column clustering in the heatmap.}

\item{sample_metadata}{A dataframe containing the metadata information for the grouping of the columns.}

\item{annotation_colors}{A list assigning choosen colors to the corresponding groups.}

\item{annotation_name_side}{The side of the column annotation description, default = "right".}

\item{show_row_names}{A Boolean switching rownames on the heatmap on and off.}

\item{show_column_names}{A Boolean switching colum names on the heatmap on and off.}

\item{row_annotation}{A Boolean switching row annotation on the heatmap on and off.}

\item{row_annotation_method}{A string setting the annotation method of the heatmap.}

\item{row_anno_names}{A list containing choosen rownames to use for the row annotation.}

\item{row_anno_number}{An integer setting the number of automatic annotations assigned per cluster.}

\item{fontsize_rowAnnotation}{An integer setting the fontsize of the optional row annotation.}

\item{fontsize_columnNames}{An integer setting the fontsize of the column names.}

\item{fontsize_rowNames}{An integer setting the fontsize of the row names.}

\item{title_heatmapLegend}{A string setting the changeable title of the legend, default "Expression".}

\item{WidthNum}{A float setting the width of the heatmap.}

\item{HeightNum}{A float setting the height of the heatmap.}

\item{UnitSize}{A string such as "cm" or "inch" to set the unit of HeightNum and WidthNum.}
}
\value{
The colors used in heatmap based on either ComplexHeatmap default or RColorBrewer.
}
\description{
Creating a color scheme based on the available color palettes of RColorBrewer for the heatmap.
}
\examples{
colorPalette <- "RdBu"
}
