% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{DGEADESeq2}
\alias{DGEADESeq2}
\title{DGEADESeq2}
\usage{
DGEADESeq2(rawCounts, metadata, grouping_columns, comparisons)
}
\arguments{
\item{rawCounts}{A matrix containing raw counts (integers).}

\item{metadata}{A dataframe containing the meta data information of the
raw counts matrix.}

\item{grouping_columns}{A list of the columns names in the metadata file that
are used to generate the groups for the DEA that are compared with each
other.}

\item{comparisons}{An element of the class list, that contains optional
information on specific comparisons that are to be conducted
(simplifies the results).}
}
\value{
A list containing the voom data, fitting of the linear model, results
of the contrasts, and the generated design.
}
\description{
Performs a differential expression analysis based on DESeq2
}
\examples{
set.seed(1)

counts <- matrix(rnbinom(80, mu = 20, size = 1), ncol = 4)
colnames(counts) <- paste0("Sample", 1:4)
rownames(counts) <- paste0("Gene", 1:20)
storage.mode(counts) <- "integer"

metadata <- data.frame(
    segment = factor(rep(c("A", "B"), each = 2)),
    row.names = colnames(counts)
)

comparisons <- list(
    Comp1 = c("A", "B")
)

results_DESeq2 <- DGEADESeq2(
    rawCounts = counts,
    metadata = metadata,
    grouping_columns = "segment",
    comparisons = comparisons
)
}
