% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{DGEAedgeR}
\alias{DGEAedgeR}
\title{Performs a differential expression analysis based on edgeR}
\usage{
DGEAedgeR(
  rawCounts,
  metadata,
  grouping_columns,
  comparisons = NULL,
  prefix = "DEA"
)
}
\arguments{
\item{rawCounts}{A matrix containing raw counts (integers).}

\item{metadata}{A dataframe containing the meta data information of the raw counts matrix.}

\item{grouping_columns}{A list of the columns names in the metadata file that are used to generate the groups for the DEA that are compared with each other.}

\item{comparisons}{An element of the class list, that contains optional information on specific comparisons that are to be conducted (simplifies the results).}

\item{prefix}{A string used as a prefix for the contrast_names that are generated.}
}
\value{
A list containing the voom data, fitting of the linear model, results of the contrasts, and the generated design.
}
\description{
Performs a differential expression analysis based on edgeR
}
