% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{DGEALimma}
\alias{DGEALimma}
\title{DGEALimma}
\usage{
DGEALimma(
  rawCounts,
  metadata,
  grouping_columns,
  comparisons = NULL,
  prefix = "DEA"
)
}
\arguments{
\item{rawCounts}{A matrix containing raw counts (integers).}

\item{metadata}{A dataframe containing the meta data information of the raw
counts matrix.}

\item{grouping_columns}{A list of the columns names in the metadata file that
are used to generate the groups for the DEA that are compared with each
other.}

\item{comparisons}{An element of the class list, that contains optional
information on specific comparisons that are to be conducted (simplifies the
results).}

\item{prefix}{A string used as a prefix for the contrast_names that are
generated.}
}
\value{
A list containing the voom data, fitting of the linear model, results
of the contrasts, and the generated design.
}
\description{
Performs a differential expression analysis based on limma voom
}
\examples{
rawCounts <- read.csv(
system.file("extdata/RawDataExamplePackageNanostring.csv",
package = "DgeaHeatmap"))
rawCounts <- build_matrix(rawCounts, 1)
metadata <- read.csv(system.file("extdata/MetaDataPackageNanostring.csv",
package = "DgeaHeatmap"))
grouping_columns <- c("segment", "region", "class", "slide_name")
comparisons <- list(Comp1 = c("Geometric_Segment_glomerulus_DKD_disease3",
"PanCK_tubule_DKD_disease4"), Comp2 = c("neg_tubule_DKD_disease4",
"PanCK_tubule_DKD_disease4"))
prefix <- "DEA"
DGEAListResults <- DGEALimma(rawCounts, metadata, grouping_columns,
comparisons, prefix)
}
