% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{draw_adv_heatmap}
\alias{draw_adv_heatmap}
\title{Draw the advanced heatmap}
\usage{
draw_adv_heatmap(
  ncounts_matrix,
  column_name = "Heatmap",
  colorPalette = NULL,
  col_ha = NULL,
  show_row_names = FALSE,
  show_column_names = TRUE,
  fontsize_title = 15,
  fontsize_columnNames = 6,
  fontsize_rowNames = 4,
  fontsize_cluster_labels = 8,
  fontsize_heatmap_legend = 10,
  fontsize_heatmap_legend_labels = 8,
  title_heatmapLegend = "Expression",
  WidthNum = 4.5,
  HeightNum = 3,
  UnitSize = "cm",
  row_annotation = FALSE,
  annotation_for_rows = NULL,
  row_split = NULL,
  col_split = NULL,
  row_dend = TRUE,
  col_dend = TRUE
)
}
\arguments{
\item{ncounts_matrix}{An input matrix to create the heatmap.}

\item{column_name}{A string to set the title of the heatmap.}

\item{colorPalette}{Name of the colorPalette used for the heatmap, default <- NULL.}

\item{col_ha}{A "HeamapAnnotation" object if "sample_metadata" is provided, otherwise NULL.}

\item{show_row_names}{A Boolean switching rownames on the heatmap on and off, default <- FALSE.}

\item{show_column_names}{A Boolean switching colum names on the heatmap on and off, default <- TRUE.}

\item{fontsize_title}{An integer setting the font size of the heatmap title, default <- 15.}

\item{fontsize_columnNames}{An integer setting the font size of the column names, default <- 6.}

\item{fontsize_rowNames}{An integer setting the font size of the row names, default <- 4.}

\item{fontsize_cluster_labels}{An integer setting the font size of the cluster labels, default <- 8.}

\item{fontsize_heatmap_legend}{An integer setting the font size of the heatmap legend title, default <- 10.}

\item{fontsize_heatmap_legend_labels}{An integer setting the font size of the heatmap legend labels, default <- 8.}

\item{title_heatmapLegend}{A string setting the changeable title of the legend, default "Expression".}

\item{WidthNum}{A float setting the width of the heatmap, default <- 4.5.}

\item{HeightNum}{A float setting the height of the heatmap, default <- 3.}

\item{UnitSize}{A string such as "cm" or "inch" to set the unit of HeightNum and WidthNum, default <- "cm".}

\item{row_annotation}{A Boolean switching row annotation on the heatmap on and off, default <- FALSE.}

\item{annotation_for_rows}{Numeric index from matrix for the row annotation}

\item{row_split}{An integer seeting the row clustering by kmeans.}

\item{col_split}{An integer seeting the column clustering by kmeans.}

\item{row_dend}{A Boolean switching row clustering by hierarchical clustering on (default).}

\item{col_dend}{A Boolean switching column clustering by hierarchical clustering on (default).}
}
\value{
Row annotation
}
\description{
Draw the advanced heatmap
}
\examples{
x <- 1
seed <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv", package = "DgeaHeatmap"))
matrixCounts <- build_matrix(input_data, x)
ncounts_matrix <- scale_counts(matrixCounts)
set.seed(seed)
row_split <- NULL
row_dend <- TRUE
col_split <- NULL
col_dend <- TRUE
groups <- c("3_DKD_glomerulus_Geometric_S", "1B_DKD_glomerulus_Geometric_S", "2B_DKD_glomerulus_WT")
sample_names <- c(colnames(ncounts_matrix))
group_assignment <- sapply(sample_names, function(sample) {
  matched <- groups[sapply(groups, function(g) grepl(g, sample))]
  if (length(matched) > 0) matched[1] else NA
})
stopifnot(length(sample_names) == length(group_assignment))
sample_metadata <- data.frame(Group = group_assignment, row.names = sample_names)
all(colnames(ncounts_matrix) == rownames(sample_metadata))
group_colors <- list(Group = c("3_DKD_glomerulus_Geometric_S" = "#1b9e77",
"1B_DKD_glomerulus_Geometric_S" = "#7570b3",
"2B_DKD_glomerulus_WT" = "#e7298a"))
col_ha <- set_sample_annotation(sample_metadata = sample_metadata, annotation_colors = group_colors)
ht <- draw_adv_heatmap(ncounts_matrix)
}
