% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{extractDEGenes}
\alias{extractDEGenes}
\title{Extracts the results of the DEA with DESeq2}
\usage{
extractDEGenes(
  results_list,
  comparisons,
  only_up = FALSE,
  only_down = FALSE,
  up_down = FALSE,
  only_sig = FALSE,
  padj_cutoff = 0.05,
  lfc_cutoff = 0
)
}
\arguments{
\item{results_list}{A list containing the results of the DEA.}

\item{comparisons}{An element of the class list, that contains optional information on specific comparisons that are to be conducted (simplifies the results).}

\item{only_up}{A Boolean opting for only up regulated genes, default = FALSE.}

\item{only_down}{A Boolean opting for only down regulated genes, default = FALSE.}

\item{up_down}{A Boolean opting for only up and down regulated genes, default = FALSE.}

\item{only_sig}{A Boolean opting for only significantly regulated genes, default = FALSE.}

\item{padj_cutoff}{A Float, setting the adjusted p-values cutoff, default = 0.05.}

\item{lfc_cutoff}{A Float, setting the fold change cutoff, default = 0.}
}
\value{
A list containing the voom data, fitting of the linear model, results of the contrasts, and the generated design.
}
\description{
Extracts the results of the DEA with DESeq2
}
