% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{function_complexHeatmap_var}
\alias{function_complexHeatmap_var}
\title{Creating a heatmap with annotation of x most variable rows(genes).}
\usage{
function_complexHeatmap_var(
  topGenes_matrix,
  probes,
  number_of_annotations_per_cluster,
  k,
  seed,
  Title,
  fontsize_columnNames,
  fontsize_rowNames,
  title_heatmapLegend,
  WidthNum,
  HeightNum,
  UnitSize
)
}
\arguments{
\item{topGenes_matrix}{An input matrix to create the heatmap.}

\item{probes}{A list of strings to summarize biological replicated in the columns.}

\item{number_of_annotations_per_cluster}{An integer used to set the number of annotations per cluster in the heatmap.}

\item{k}{An integer used to set number of cluster in the heatmap.}

\item{seed}{An integer used to set seed for reproducibility and k-Mean clustering.}

\item{Title}{A string to set the title of the heatmap.}

\item{fontsize_columnNames}{An integer used to set the font size of the columns in the heatmap.}

\item{fontsize_rowNames}{An integer used to set the font size of the rownames in the heatmap.}

\item{title_heatmapLegend}{A string setting the title of the legend of the heatmap.}

\item{WidthNum}{A float setting the width of the heatmap.}

\item{HeightNum}{A float setting the height of the heatmap.}

\item{UnitSize}{A string such as "cm" or "inch" to set the unit of HeightNum and WidthNum.}
}
\value{
A plotted heatmap of the input data.
}
\description{
Creating a heatmap with annotation of x most variable rows(genes).
}
\examples{
topGenes_matrix <- matrix
probes <- list("Female", "Male")
number_of_annotations_per_cluster <- 10
k <- 3
seed <- 1
Title <- "Heatmap with annotation of most variable genes"
}
