% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{genRawReadCountTable}
\alias{genRawReadCountTable}
\title{Function to generating a raw read count table.}
\usage{
genRawReadCountTable(demoData)
}
\arguments{
\item{demoData}{A demoData object, storing all expression, annotation, and probe information linked together.}
}
\value{
A dataframe containing only the raw read counts that have passed quality control.
}
\description{
Function to generating a raw read count table.
}
\examples{
demoData <- file
}
