% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{genRawReadCountTable}
\alias{genRawReadCountTable}
\title{Function to generating a raw read count table.}
\usage{
genRawReadCountTable(demoData)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object, storing all expression, annotation, and probe information linked together.}
}
\value{
A dataframe containing only the raw read counts that have passed quality control.
}
\description{
Function to generating a raw read count table.
}
\examples{
datadir <- system.file("extdata", "WTA_NGS_Example", package="GeoMxWorkflows")
DCCFiles <- dir(file.path(datadir, "dccs"), pattern = ".dcc$", full.names = TRUE, recursive = TRUE)
PKCFiles <- unzip(zipfile = dir(file.path(datadir, "pkcs"), pattern = ".zip$", full.names = TRUE, recursive = TRUE))
SampleAnnotationFile <- dir(file.path(datadir, "annotation"), pattern = ".xlsx$", full.names = TRUE, recursive = TRUE)
rawDataObject <- suppressWarnings(GeomxTools::readNanoStringGeoMxSet(dccFiles = DCCFiles, pkcFiles = PKCFiles, phenoDataFile = SampleAnnotationFile, phenoDataSheet = "Template", phenoDataDccColName = "Sample_ID", protocolDataColNames = c("aoi","roi"), configFile = NULL, analyte = "RNA", phenoDataColPrefix = "", experimentDataColNames = NULL))
demoData <- add_demoElem(rawDataObject)
NanoStringNCTools::assayDataApply(demoData, MARGIN=1, FUN=mean, elt="demoElem")[seq_len(5)]
VGroup <-"aoi"
vElt <- "demoElem"
demoData <- split_data_by_column(demoData, VGroup, vElt)
QCPassed <- aExprsDataQC(demoData, "QCFlags")
df_Exp <- genRawReadCountTable(demoData)
demoData <- file
}
