% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{Kmean_generation}
\alias{Kmean_generation}
\title{Generates K-means for the columns and rows of a matrix.}
\usage{
Kmean_generation(m_top_genes_matrix, k)
}
\arguments{
\item{m_top_genes_matrix}{An input matrix for which they K-Means
are calculated next.}

\item{k}{Number k of how many clusters are created.}
}
\value{
All calculated K-Means of the matrix.
}
\description{
Generates K-means for the columns and rows of a matrix.
}
\examples{
x <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
    package = "DgeaHeatmap"
))
matrixCounts <- build_matrix(input_data, x)
m_top_genes_matrix <- scale_counts(matrixCounts)
seed <- 1
k <- 1
set.seed(seed)
k_means <- Kmean_generation(m_top_genes_matrix, k)
}
