% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{prepare_dge_list}
\alias{prepare_dge_list}
\title{DGEAedgeR}
\usage{
prepare_dge_list(rawCounts, metadata)
}
\arguments{
\item{rawCounts}{A matrix containing raw counts (integers).}

\item{metadata}{A dataframe containing the meta data information of the raw
counts matrix, with a column containing the comparison factors.}
}
\value{
A DGEList object.
}
\description{
Creates a contrast matrix for edgeR
}
\examples{
rawCounts <- read.csv(system.file(
  "extdata/RawDataExamplePackageNanostring.csv",
  package = "DgeaHeatmap"
))
rawCounts <- build_matrix(rawCounts, 1)
metadata <- read.csv(system.file(
  "extdata/MetaDataPackageNanostring.csv",
  package = "DgeaHeatmap"
))
y <- prepare_dge_list(rawCounts, metadata)
}
