% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{print_heatmap}
\alias{print_heatmap}
\title{Function to build a heatmap using other functions.}
\usage{
print_heatmap(
  m_top_genes_matrix,
  title,
  split,
  anno,
  fontsize_columnNames,
  fontsize_rowNames,
  title_heatmapLegend,
  WidthNum,
  HeightNum,
  UnitSize,
  color_Palette
)
}
\arguments{
\item{m_top_genes_matrix}{An input matrix for which the heatmap is created.}

\item{title}{A string used to set the title of the heatmap.}

\item{split}{A dataframe containing information about the clustering of
the rows of the input matrix.}

\item{anno}{A numeric index containing row informations for annotation.}

\item{fontsize_columnNames}{An integer used to set the font size of the
columns in the heatmap.}

\item{fontsize_rowNames}{An integer used to set the font size of the
rownames in the heatmap.}

\item{title_heatmapLegend}{A string setting the title of the
legend of the heatmap.}

\item{WidthNum}{A float setting the width of the heatmap.}

\item{HeightNum}{A float setting the height of the heatmap.}

\item{UnitSize}{A string such as "cm" or "inch" to set the unit of
HeightNum and WidthNum.}

\item{color_Palette}{Name of the colorPalette used for the heatmap.}
}
\value{
A plotted heatmap.
}
\description{
Function to build a heatmap using other functions.
}
\examples{
x <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
package = "DgeaHeatmap"))
matrixCounts <- build_matrix(input_data, x)
m_top_genes_matrix <- scale_counts(matrixCounts)
k <- 1
split <- performing_kMeans(m_top_genes_matrix, k)
seed <- 1
fontsize_rowAnnotation <- 8
set.seed(seed)
m_kmeans <- Kmean_generation(m_top_genes_matrix, k)
number_of_annotations_per_cluster <- 5
top_x_genes_cluster <- most_variable_genes(m_kmeans,
number_of_annotations_per_cluster, k)
anno <- set_annotation(m_top_genes_matrix, top_x_genes_cluster,
fontsize_rowAnnotation)
title <- "Heatmap of Data"
fontsize_columnNames <- 6
fontsize_rowNames <- 4
title_heatmapLegend <- "Expression"
WidthNum <- 4.5
HeightNum <- 3
UnitSize <- "cm"
color_Palette <- "RdBu"
set.seed(seed)
print_heatmap(m_top_genes_matrix, title, split, anno, fontsize_columnNames,
fontsize_rowNames, title_heatmapLegend, WidthNum, HeightNum,
UnitSize, color_Palette)
}
