% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_building.R
\name{scale_counts}
\alias{scale_counts}
\title{Function to Z-count scale the values of a matrix.}
\usage{
scale_counts(countsmatrix)
}
\arguments{
\item{countsmatrix}{An input matrix whose values are scaled by Z-count.}
}
\value{
A new matrix with Z-count scaled values.
}
\description{
Function to Z-count scale the values of a matrix.
}
\examples{
x <- 1
input_data <- read.csv(system.file("extdata/testfile_counts.csv",
    package = "DgeaHeatmap"
))
countsmatrix <- build_matrix(input_data, x)
scaled_counts <- scale_counts(countsmatrix)
}
