% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{split_data_by_column}
\alias{split_data_by_column}
\title{Splitting data by group column with feature, pheno or protocol data to then get the mean.}
\usage{
split_data_by_column(demoData, vGroup, vElt)
}
\arguments{
\item{demoData}{A demoData object, storing all expression, annotation, and probe information linked together.}

\item{vGroup}{Group of features in column.}

\item{vElt}{Character string of the expression matrix name.}
}
\value{
A demoData object.
}
\description{
Splitting data by group column with feature, pheno or protocol data to then get the mean.
}
\examples{
demoData <- file
VGroup <-"aoi"
elt <- "demoElem"
}
