% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBuildingNanostringGeoMxDSP.R
\name{split_data_by_column}
\alias{split_data_by_column}
\title{Splitting data by group column with feature,
pheno or protocol data to then get the mean.}
\usage{
split_data_by_column(demoData, vGroup, vElt)
}
\arguments{
\item{demoData}{A NanostringGeoMxSet object,
storing all expression, annotation, and probe information linked together.}

\item{vGroup}{Group of features in column.}

\item{vElt}{Character string of the expression matrix name.}
}
\value{
A NanostringGeoMxSet object.
}
\description{
Splitting data by group column with feature,
pheno or protocol data to then get the mean.
}
\examples{
datadir <- system.file("extdata", "WTA_NGS_Example",
  package = "GeoMxWorkflows"
)
DCCFiles <- dir(file.path(datadir, "dccs"),
  pattern = ".dcc$", full.names = TRUE, recursive = TRUE
)
PKCFiles <- unzip(
  zipfile = dir(file.path(datadir, "pkcs"),
    pattern = ".zip$", full.names = TRUE, recursive = TRUE
  ),
  exdir = tempdir()
)
SampleAnnotationFile <- dir(file.path(datadir, "annotation"),
  pattern = ".xlsx$", full.names = TRUE, recursive = TRUE
)
rawDataObject <- suppressWarnings(
  GeomxTools::readNanoStringGeoMxSet(
    dccFiles = DCCFiles,
    pkcFiles = PKCFiles, phenoDataFile = SampleAnnotationFile,
    phenoDataSheet = "Template", phenoDataDccColName = "Sample_ID",
    protocolDataColNames = c("aoi", "roi"), configFile = NULL,
    analyte = "RNA", phenoDataColPrefix = "", experimentDataColNames = NULL
  )
)
demoData <- add_demoElem(rawDataObject)
NanoStringNCTools::assayDataApply(
  demoData,
  MARGIN = 1, FUN = mean, elt = "demoElem"
)[seq_len(5)]
VGroup <- "aoi"
vElt <- "demoElem"
demoData <- split_data_by_column(demoData, VGroup, vElt)
}
