% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_functions.R
\name{summarize_edgeR_DEA}
\alias{summarize_edgeR_DEA}
\title{Summarize the results of the DEA with edgeR}
\usage{
summarize_edgeR_DEA(
  results_edgeR,
  lfc_threshold = 1,
  fdr_threshold = 0.05,
  return_significant_genes = TRUE
)
}
\arguments{
\item{results_edgeR}{A list containing the results of the DEA.}

\item{lfc_threshold}{A Float, setting threshold value for the fold change.}

\item{fdr_threshold}{A Float, setting the False Discovery Rate threshold.}

\item{return_significant_genes}{A Boolean opting for the output of
significantly regulated genes.}
}
\value{
A list containing the voom data, fitting of the linear model, results
of the contrasts, and the generated design.
}
\description{
Summarize the results of the DEA with edgeR
}
\examples{
results_edgeR <- readRDS(
    system.file(
        "extdata",
        "mini_edgeR_results.rds",
        package = "DgeaHeatmap"
    )
)

summarize_edgeR_DEA(results_edgeR)
}
